Change GAME_LOOP_TICK_RATE to actually represent a rate Game loop nows ticks 30 times per second instead of 1000 / 30. This may make some per second calculations a bit smoother.

+++ GameServer/ECS-Services/GameServiceUtils.cs
@@ -8,7 +8,7 @@ namespace DOL.GS
     public static class GameServiceUtils
     {
         private static readonly Logger log = LoggerManager.Create(MethodBase.GetCurrentMethod().DeclaringType);
-        private static long HalfTickDuration => GameLoop.TickDuration / 2;
+        private static double HalfTickDuration => GameLoop.TickDuration / 2;
 
         public static bool ShouldTick(long tickTime)
         {
